// src/app/account-management/Models/AccountModels.ts
/**================================== NOTE ===============================================
 * This section to be modify at actual implementation. Thais is just an example!!
 * --------------------------------------------------------------------------------------**/

export class AccountTest {
  id: string;
  accountID: string;
  accountNumber: string;
  name: string;
  statusId: number;
  address: string;
  phone: string;
}